werkplaats-3-rest-coding-cats-main/aanwezigheids-tool/requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.0.9
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap-datetimepicker==1.5
django-bootstrap5==22.2
django-crispy-forms==2.0
django-filter==23.1
django-tables2==2.5.3
docopt==0.6.2
Faker==16.6.1
idna==3.4
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
pipreqs==0.4.11
platformdirs==2.6.2
pydantic==1.10.6
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.4.0
yarg==0.1.9
